Expose refresh rates to RefreshRateController
To support use cases beyond throttle/no-throttle RefreshRateController
needs to have visibility into the refresh rates that the underlying
panel can switch to without a modeset (seamlessly).

This change adds functionality for RefreshRateController to query this.
In a follow up CL, the list of supported modes will be filtered to those
that can only be switched to seamlessly.

The results of the query for supported refresh rates is not currently
used, and it will be used in a follow up CL.

BUG=b:323362145
TEST=RefreshRateControllerTest passes

Change-Id: Id5031cadf2a63bdb0ccbd7697c7b45fe69e230e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5293073
Commit-Queue: Drew Davenport <ddavenport@chromium.org>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Chris Bookholt <bookholt@chromium.org>
Reviewed-by: Andrew Wolfers <aswolfers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1266631}