Mahi: Add use-fake-mahi-manager to facilitate UI development
When passing in the switch --use-fake-mahi-manager, the UI will use the
fake manager instead of the real one. This helps us with the UI
development and decouple the UI from the backend work.

This CL also moves FakeMahiManager to //ash, since this class will now
only be used by //ash and also use some capabilities in //ash.

Bug: b:327309809
Change-Id: If378ec8edb69c78bbf2725b3b5ad61691425ff03
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5332295
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Andre Le <leandre@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1266653}