vt: Fix resource id sequencing cross-process
Resource IDs for view transition snapshots are assigned using a
base::AtomicSequence in Blink. However, this uses a process-static
counter. When performing a view transition to a prerendred page, the two
pages are in separate processes so the ids generated for new snapshots
may collide with old snapshot ids.

This CL makes three conceptual changes:

First we pass the next_id from the old document to the new one in the
view transition state so that the id sequence can be continued in the
new document, even if it's in a separate process.

Second, the id sequence is moved onto the per-Document
ViewTransitionSupplement to ensure there isn't any confusion between
transitions occurring in separate documents being rendered in the same
process.

Last, the ViewTransitionElementResourceID type is turned into a tuple
consisting of a per-transition UnguessableToken "transition id" and a
per-snapshot "id". The transition id is preserved across the old and new
Document's ViewTransition object by being passed through the
ViewTransitionState. The individual snapshot "id"s are guaranteed to be
unique only within a specific "transition id". This ensures snapshot ids
from separate transitions being applied within the same compositor
instance cannot colide.

Note: The "navigation id" of a capture request and used in viz to allow
sharing resources between separate documents when they're transitioning
is a separate identifier but requires only that the id be the same for
old and new captures and unique across transitions. The "transition id"
has these properties so is used to set the navigation id in this case.
However, unlike the transition id, for same-document navigations it
continues to be an empty id.

Fixed: 41496294
Change-Id: I3d50fb6c91ee844ab1506217f921170eefd0bcde
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5314213
Reviewed-by: danakj <danakj@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1266655}