FLEDGE: Multi-bid: fix ordering of semantic checks
... and switch the length check to the simpler-to-describe length of
the sequence (rather than number of actual bids in the sequence).

The core of the change, however, is making sure that the checks
corresponding to the conversion to the IDL Sequence<GenerateBidOutput>
type happen before looking at the semantics of entries, as that's how
things are supposed to be spec'd to work.

A few spots that were ignoring the passed in error_prefix and
hardcoding their own were also fixed.

Bug: 323856489
Change-Id: I9ff619e7481b2350c677df96a500a701de157582
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5325074
Reviewed-by: mmenke <mmenke@chromium.org>
Commit-Queue: Maks Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1266682}