[crypto] Fix UnexportableKeyMac bug
UnexportableKeyMac relies on SecKeyCopyAttributes returning the
SecKeyRef along the rest of the key attributes, which was the behaviour
of the test fake. Unfortunately, the real API does not do this, so the
code does not work against the real keychain. Fix this.

Bug: 326451282
Change-Id: I56fcd9e34ec73771a16e00a370e5c7aa771a6cd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5332152
Reviewed-by: Adam Langley <agl@chromium.org>
Commit-Queue: Adam Langley <agl@chromium.org>
Auto-Submit: Nina Satragno <nsatragno@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1266702}