Remove tools/perf third party mock usage
Removes usage of Telemetry's third party copy of mock in //tools/perf
in favor of the built-in version that has been available since the
switch to Python 3.

Since this change is dependent on the corresponding one in Catapult in
order for the .pydeps file to update accordingly, also manually rolls
Catapult to pick up the change.

Bug: 324114898
Change-Id: I3d3c7bb82fdc35648a2144b1b9f699234308ef73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5332640
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Wenbin Zhang <wenbinzhang@google.com>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1266710}