[Start Surface] Remove ability to create GTS
Remove the ability of start surface to create a grid tab switcher.

This is an incomplete cleanup, but it does remove the calls to
`createGridTabSwitcher` which is important so that we can more
easily change the list of parameters pass into the
TabSwitcherCoordinator constructor.

Bug: 327429875
Change-Id: I4814d33003671d3a5006e4cb51c3d8eecf73c268
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5331454
Reviewed-by: Xi Han <hanxi@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1266718}