Remove the expired flag ash-overview-button
This flag was introduced and run as an experiment to see whether it
will help improve the discoverability of the overview mode, which
should be fine to be removed now after launching the bento button.
A new flag can be added with new requirement, e.g., no experiment
related logic.

Bug: b/40257985
Change-Id: Ifa2131c1d12dc96bc02645b2579518fe5d56c81e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5333353
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Commit-Queue: Min Chen <minch@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1266727}