Shared Storage: Correct expected iterator histogram count in perf tests
We had an off-by-one error in the expected histogram sample counts
for `keys()` and `entries()` stories, because we forgot about the
final call to `blink::SharedStorage::GetNextIterationResult()` that
initiates a call to `blink::SharedStorage::MakeEndOfIteration()` to
terminate the iteration.

We correct the expected counts.

Bug: 1516507,1218540
Change-Id: Ie9383b43b7c1f344c56a3d8be0f178d1711ff8fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5331461
Commit-Queue: Yao Xiao <yaoxia@chromium.org>
Reviewed-by: Yao Xiao <yaoxia@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1266729}