[DBSC] Extract a base class from CanonicalCookie
This pulls CanonicalCookie apart into a base class, CookieBase. Most of
the data fields are pulled into CookieBase, along with the methods that
check inclusion on cookie read/write. CanonicalCookie retains some data
members that are not needed in the base class, and methods that allow
creation of an actual CanonicalCookie. The base class has its
constructors protected. A couple of small things are pulled out of
canonical_cookie.{h,cc} to allow use in the base class.

The base class will be used to implement DBSC.

Bug: 324405105, 41495201
Change-Id: I0c2249da2f95c4756564d1f4c1e8025d03e0fe75
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5277605
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Lily Chen <chlily@chromium.org>
Reviewed-by: Steven Bingler <bingler@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1266732}