pine: Migrate unit tests from full restore service p2
Migrates the rest of the FullRestoreServiceTest tests. Disables forest
feature for the non-forest tests so that the pass if forest is enabled.

The non-forest tests simulate clicks on the notification to check
cancelling works. We cannot simulate clicks on the dialog with the unit
tests so add a browser test for that.

Test: unit_tests *FullRestoreService*
Bug: b/326982900
Change-Id: I81465818b284e50bc01057f73ddc65221b3083d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5328433
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1266734}