Tab Organization: Create actions component
Allows the "Create group" button and other actions to be reused in
different places for the single-group UI and multi-group experiment. The
former has per-group actions, while the latter has a single set of
actions for all groups.

Implementing the multi-group behavior will be done in a separate CL.

Bug: b/324942554
Change-Id: Ia67af1377311f7548fe447b90b7adfa0de847302
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5328614
Reviewed-by: Shibalik Mohapatra <shibalik@chromium.org>
Reviewed-by: Darryl James <dljames@chromium.org>
Reviewed-by: Alison Gale <agale@chromium.org>
Commit-Queue: Emily Shack <emshack@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1266754}