utr: Set swarming env-var in recipe invocations
Some parts of the recipe need the swarming server env var set to
control what it does. (eg: what cas instance to upload test inputs to)

So this just hardcodes the chromium-swarm server for now. When
support for internal builders is added, it can do something a
little smarter to use the internal server.

Bug: 41492688
Change-Id: I9ce215fce712c5453e8176a960933c54fcf8e2a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5327502
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1266761}