Move functions from DataExchangeDelegate to SecurityDelegate
This CL is a refactor only to move code.  There is no change of logic.

A separate SecurityDelegate is created for each VM, and potentially for
other clients such as lacros and arc. By moving the file sharing and
translation functions to SecurityDelegate we can allow each VM to do the
different path translations required.

Bug: b/323465931
Change-Id: I275ddf4848a318ac1bc1bfbf9538385619b18736
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5290176
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1266768}