Layout: Fix a crash in AppendText()
When we update the content of a Text node, we do:
1. Update Text::data_
2. Call InlineNode::SetTextWithOffset()
3. Update LayoutText::text_ with transformation

In SetTextWithOffset(), we call InlineNodeDataEditor::Prepare(), which
might build an OffsetMapping with InlineItemsBuilderTemplate.  It
expects that LayoutText::TransformedText() is based on the text before
step 1.

So, we must not transform OriginalText() on the fly in
InlineItemsBuilderTemplate because OriginalText() is Text::data_. We
need to record TextOffsetMap which was generated for LayoutText::text_.

This CL introduces a new struct `VariableLengthTransformResult`, which
contains a TextOffsetMap instance and an original text length.  It's
per-LayoutText object, but is stored in LayoutView because storing it in
each of LayoutText increases memory consumption.

Bug: 40532763, 325306591
Change-Id: I63ec8275ec4454ec6a7d8f3864b6f6597e667dd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5320433
Commit-Queue: Kent Tamura <tkent@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Auto-Submit: Kent Tamura <tkent@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1266856}