[CompositeClipPathAnimation] Make animation updates update paint props
During edge experimentation, it was discovered that occasionally a crash would result when repainting, if an element had been set with mask clip: true for a composited clip path animation but had since become non compositable. This can happen if an existing animation is updated to have properties that make it non-compositable (eg. positive delay)

This is related to crbug.com/1497795 which has the exact same failure trace.

Bug: 686074

Change-Id: I052053c87bab9b70d3224a555150a92976a9cc55
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5218911
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Commit-Queue: Claire Chambers <clchambers@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1266880}