Reland #3: "[A11y] Stable ids for AXObjects with DOM nodes"
This is a reland of commit 01466963a42b3bd5fdbe5bc5ac17bff6304b058d
Reland #1 was reverted because TAST test inputs.VirtualKeyboardAccent
still fails sometimes.
Reland #2 was reverted because of failing mac test:
DumpAccessibilityEventsTest.AccessibilityEventsAriaComboBoxNext
https://ci.chromium.org/ui/p/chromium/builders/ci/mac11-arm64-rel-tests/34319/overview
https://ci.chromium.org/ui/p/chromium/builders/ci/mac13-arm64-rel-tests/17422/overview
This occurred because we accidentally picked up a
flaky test result via rebase_dump_accessibility_tree_test.py.

Original change's description:
Reland: [A11y] Stable ids for AXObjects with DOM nodes

Reland of CL:4027071. The revert was caused because of the breaking
of a TAST test. Basically, some code in AXTreeSerializer that deleted
subtrees of old child ids is not triggered as often, because if the
DOM node doesn't change, the child ID doesn't. That's a good thing
usually, but if the node is reattached with new style, we need to
make sure its children are reconsidered. The fix is that in the
specific case where a node is reattached, call ChildrenChanged() instead
of just marking it dirty -- in
AXObjectCacheImpl::UpdateCacheAfterNodeIsAttached().

----

Use the DOM node for the AXID for any AXObject with a DOM node.
Other AXObjects will still generate their AXID, but in their own
numerical namespace. For now, any AXID < 0 will be a generated ID.

Role changes and alerts required some new browser-side code, because
rather than the object being destroyed and created with a new id,
the ID now stays the same.

In some cases, test results have improved, e.g. rather than an object
being destroyed and created with a different ID, it retains the same ID
and a change event is fired for it.

Benefits:
- Removes over 100 lines of implementation code
- Use less memory: removes ids_in_use_ (set) and
node_object_mapping_ (map).
- Reduce map lookups when an AXObject is removed -- based on the ID
value, we can determine whether it has a DOM node and if it doesn't,
we already know it cannot be in some of the maps
- Make it easier for screen readers to maintain the user's point of
regard within content, even when the layout or role changes. Users
sometimes lose their place, as the screen reader may move the user
to the top. A hope is that this will reduce the occurrences of that
happening. It may even be possible for screen readers to keep their
place in tabs where the AX tree is released from memory and later
regenerated, which could open up possibilities to using less memory.
- Enable the follow-ups listed below.

Follow-up work:
- Refactor content-visibility: auto change handling in a11y
(crbug.com/1380449).
- See if the dom_node_id field in AXNodeData is needed anymore now
that any id >0 is the same as the dom node id.
- Look at simplifying or removing reparenting computations in
AXTreeSerializer. Maybe we can get rid of expensive
AnyDescendantIsReparented().
- Merge GetAXID() and GetExistingAXID() -- see TODO.

Bug: none
Change-Id: I20e35548e249e3675bacf9c97cdb309ceb396c98
Cq-Include-Trybots: luci.chrome.try:chromeos-betty-pi-arc-chrome-accessibility-fyi,chromeos-eve-chrome
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5328829
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: David Tseng <dtseng@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1266923}