Force tighter check on extension/process id.
This CL tightens up a check that is commented to enforce that a
cross-renderer resource load request is being issued by a process that
belongs to a particular extension. But the existing check appears to
only require that the request comes from a process that is an extension,
but not necessarily the extension in question.

Change-Id: I0174f99c1568e1ecee12334a3e0e3ae94ce51b60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5324460
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: W. James Maclean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1266943}