[iOS] Make BlockedPopupTabHelper Inherit LazyWebStateUserData
This CL refactors BlockedPopupTabHelper to use LazyWebStateUserData
instead of WebStateUserData. This also makes use of the new method
GetOrCreateForWebState which is used as a replacement to FromWebState
or CreateWebState. The main difference is that when
kEnableStartupImprovements is enabled, BlockedPopupTabHelper isn't
attached in AttachTabHelpers and is instead attached when a WebState
needs to add it as a dependency.

Bug: 327166055
Change-Id: I65a91ccc1db7f3f322a007807eedcff14fc1d959
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5332411
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1266957}