[UPMLocalPwd] Allow calls to the account store during migration
Because the PasswordStore is a profile keyed service, it has to be
created from the first time it's requested. During the local passwords
migration (which also enables the split between account and profile store), both stores will exist, but the expectation was that
the account store wouldn't be invoked. However, there are no checks
in the callers of the store that explicitly verify if the split has
completed. Instead, the account store should allow certain calls, but react as it usually does when the user is not syncing, which is to act
empty.

Note: The migration is only started for local passwords, which means
the user isn't syncing.

Bug:b/326978179

Change-Id: I5425a619821cf12b856e42fee3a44c3a48f012cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5331722
Reviewed-by: Viktor Semeniuk <vsemeniuk@google.com>
Commit-Queue: Ioana Pandele <ioanap@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1266960}