[waffle]Enable the TemplateURLServiceBrowserTest.LoadKeywordData test
The test is flaky because it is we check the value of the search
engine id after "PRE" without waiting for the search engine to load.

The fix waits for the expected search engine to be loaded as
default before continuing with the test.

Fixed: b:326580767, b:41493716
Change-Id: I520396ccf9feb684f2a59c25b3fbe5e47671dbf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5331679
Auto-Submit: Jack Yammine <jyammine@google.com>
Reviewed-by: Nicolas Dossou-Gbété <dgn@chromium.org>
Commit-Queue: Nicolas Dossou-Gbété <dgn@chromium.org>
Commit-Queue: Jack Yammine <jyammine@google.com>
Cr-Commit-Position: refs/heads/main@{#1266979}