Make CSSMathExpressionOperation::HasPercentage not count anchor queries and calc-size() keywords.
This accounts for the possibility that kCalcLengthFunction (previously
known as kCalcPercentLength) does not signal the presence of
percentages.  In particular, the category can be kCalcLengthFunction
because of mixing of lengths and percentages, because of anchor queries,
and because of calc-size() keywords.

This also adds comments explaining the difference between HasPercentage
and InvolvesLayout (previously InvolvesPercentage).

This also fixes HasPercentage to propagate the percentage only from the
basis argument of the calc-size() and not from the calculation argument.

This fixes problems for calc-size() animations, which could end up
including percentages when they should not, which would break layouts
when there was no value for the percentages to resolve against.

Bug: 1309178, 313072
Change-Id: I239a8c935e74c1c53d125c2291d87934562903af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5331184
Reviewed-by: Daniil Sakhapov <sakhapov@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1266986}