[Extensions] Set crash keys for multi worker scenario.
These crash keys will be set when there are two workers added to
WorkerIdSet. It will provide extra data for the DUMP_WILL_BE_CHECK() we
perform when there are two workers.

No testing is performed since we do not know when this scenario happens,
but hope to find out with these crash keys.

Bug: 1493391
Change-Id: I5846972af146aa92c4947ed5527c175d2fedcdbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5325520
Commit-Queue: Justin Lulejian <jlulejian@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1267025}