[iOS][Spotlight] Prevent reentry to clearAndReindexIfNeeded.
If this method is called twice in quick succession, the completion block
will be called twice but the second time it's called the observers are
no longer removed so they shouldn't be readded.

I have a better long-term fix in mind but it's very disruptive and this
one needs to be cherry-picked so I kept it small.

Fixed: 41496775
Change-Id: Ie9cb761d50c909e7c652a281154ef3e6a6e40316
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5335223
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Ameur Hosni <ameurhosni@google.com>
Auto-Submit: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Ameur Hosni <ameurhosni@google.com>
Cr-Commit-Position: refs/heads/main@{#1267026}