Update System Profile in crashpad
This change makes Chrome's metrics service update the System Profile
stored in crashpad when an updated version is created.

Currently, only the first version is attached to crashpad, resulting in
fields that take longer to collect (e.g., antivirus data, which is
collected in a separate process) frequently being missing from crash
reports. This was done due to a limitation in the Crashpad API, which
has been updated as of crrev.com/c/5260024.

With this change, later updates to the SystemProfile, e.g., the one by
`metrics::MetricsService::OnFinalLogInfoCollectionDone`, which happens
after antivirus data and other fields are collected, will also be
reflected in crash reports.

Bug: b/308999458
Change-Id: I2136d420315bf0cfc516aaae7d648a130346d2c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5297579
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Commit-Queue: Jesse McKenna <jessemckenna@google.com>
Cr-Commit-Position: refs/heads/main@{#1267054}