[iOS CP] Entrypoint configuration and browser agent <-> UI communication
Added an entrypoint configuration class (which will hold more things)
in the near future, and made the flow of information between the UI and
browser agent. Fired the proper commands when user switches tabs and
the "force contextual panel entrypoint" feature is activated, with a
test config.

Bug: b/323343895, b/316372223
Change-Id: Ic93d29bd81e86a4de435986fd1cd5edd3d0cb0a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5327727
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Nicolas MacBeth <nicolasmacbeth@google.com>
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/main@{#1267072}