[Reauth] Fix crash for mandatory reauth switch
Can reproduce if clicking the toggle multiple times in a very short
time (0.5s) before the device unlock is triggered. So when this check
happens, the switch UI's value may be same as the user preference's
value if the switch is being tapped multiple times.

A straightforward fix is to simply remove this CHECK as this is not
relevant: We record the original user preference value and do a device
unlock and update the toggle to the most up-to-date value, so the UI
state during this process does not really matter.

Bug: b/325608728
Change-Id: I3524491581c6d27409a801ea3113348885bad2c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5333563
Commit-Queue: Siyu An <siyua@chromium.org>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Vidhan Jain <vidhanj@google.com>
Cr-Commit-Position: refs/heads/main@{#1267105}